+++ CHANGELOG.md
@@ -10,9 +10,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ### Added
 
+- Add `AsyncRAGSystem` and `AsyncNoEncodeRAGSystem` (#392)
 - Validate installed version of bridge framework to confirm compatibility based on `BridgeMetadata` (#323)
 - Add `NoEncodeRAGSystem` (#386)
 - Add `NoEncodeBaseKnowledgeStore` and `NoEncodeAsyncBaseKnowledgeStore` (#385)
+
+### Changed
+
+- Modified private API for `_RAGSystem`, `_AsyncRAGSystem` and no-encode versions (#392)
 
 ## [0.0.23] - 2025-05-27